Add CompanyForm and update CompaniesList layout
Introduced a new form, CompanyForm, for user to create a company. Also, the layout and functionalities of CompaniesList have been improved. Now it includes a header and footer with the new CompanyForm included under a collapsible button. Minor updates to ListItem were made to improve readability and UI consistency. Logging was also added in companiesManager.js for debug purposes.